Ajuste

+++ CalendarSchedule.API/Controllers/ScheduleController.cs
@@ -14,44 +14,31 @@ public class ScheduleController(IScheduleService _scheduleService) : ControllerB
 {
     [HttpGet]
     [Route("All")]
-    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(ScheduleDto[]))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Result))]
+    [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(PagedResult<ScheduleDto>))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     public async Task<IActionResult> GetAll([FromQuery] int page = 1, [FromQuery] int size = 10, [FromQuery] string search = "")
     {
         var scheduleDto = await _scheduleService.GetAll(page, size, search);
         if (scheduleDto.IsFailure)
             return NotFound(scheduleDto.Error);
 
-        var totalSchedule = await _scheduleService.TotalSchedules(search);
-        if (totalSchedule.IsFailure)
-            return NotFound(totalSchedule.Error);
-
-        decimal totalData = totalSchedule.Value;
-        decimal totalPage = (totalData / size) <= 0 ? 1 : Math.Ceiling((totalData / size));
-
-        if (size == 1)
-            totalPage = totalData;
-
         var schedule = scheduleDto.Value;
 
-        return Ok(new
-        {
-            totalData,
-            page,
-            totalPage,
-            size,
-            schedule
-        });
-
+        return Ok(schedule);
     }
 
     [HttpGet("{id:int}", Name = "GetScheduleId")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(ScheduleDto))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Result))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
     public async Task<IActionResult> GetById(int id)
     {
+        if (id <= 0)
+        {
+            var error = Result.Failure(Error.BadRequest($"Id ({id}) inválido"));
+            return BadRequest(error);
+        }
+
         var scheduleDto = await _scheduleService.GetById(id);
         if (scheduleDto.IsFailure)
             return NotFound(scheduleDto.Error);
@@ -62,14 +49,31 @@ public class ScheduleController(IScheduleService _scheduleService) : ControllerB
     [Authorize]
     [HttpPost]
     [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(ScheduleDto))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Result))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(Error))]
     public async Task<IActionResult> Post([FromBody] ScheduleCreateDto scheduleCreateDto)
     {
+        if (!ModelState.IsValid)
+        {
+            var errorMessage = ErroMoldeState();
+
+            var error = Result.Failure(Error.BadRequest($"Erro de validação no objeto ({nameof(ScheduleCreateDto)}): {errorMessage}"));
+            return BadRequest(error);
+        }
+
         var scheduleDto = await _scheduleService.Create(scheduleCreateDto);
         if (scheduleDto.IsFailure)
-            return NotFound(scheduleDto.Error);
+        {
+            return scheduleDto.Error.Code switch
+            {
+                "NotFound" => NotFound(scheduleDto.Error),
+                "BadRequest" => BadRequest(scheduleDto.Error),
+                "Internal" => StatusCode(StatusCodes.Status500InternalServerError, scheduleDto.Error),
+                _ => StatusCode(StatusCodes.Status500InternalServerError, scheduleDto.Error) // fallback para erro desconhecido
+            };
+        }
 
         var createResult = scheduleDto.Value;
 
@@ -79,37 +83,81 @@ public class ScheduleController(IScheduleService _scheduleService) : ControllerB
     [Authorize]
     [HttpPut("{id:int}")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(ScheduleDto))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Result))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(Error))]
     public async Task<IActionResult> Put(int id, [FromBody] ScheduleCreateDto scheduleCreateDto)
     {
         if (id <= 0)
-            return BadRequest("Dados inválidos");
+        {
+            var error = Result.Failure(Error.BadRequest($"Id ({id}) inválido"));
+            return BadRequest(error);
+        }
+
+        if (!ModelState.IsValid)
+        {
+            var errorMessage = ErroMoldeState();
+            var error = Result.Failure(Error.BadRequest($"Erro de validação do objeto ({nameof(ScheduleCreateDto)}): {errorMessage}"));
+            return BadRequest(error);
+        }
+
+        if (id != scheduleCreateDto.Id)
+        {
+            var error = Result.Failure(Error.BadRequest($"Id ({id}) da rota diferente do Id ({scheduleCreateDto.Id}) objeto"));
+            return BadRequest(error);
+        }
 
         var scheduleDto = await _scheduleService.Update(scheduleCreateDto);
         if (scheduleDto.IsFailure)
-            return NotFound(scheduleDto.Error);
+        {
+            return scheduleDto.Error.Code switch
+            {
+                "NotFound" => NotFound(scheduleDto.Error),
+                "BadRequest" => BadRequest(scheduleDto.Error),
+                "Internal" => StatusCode(StatusCodes.Status500InternalServerError, scheduleDto.Error),
+                _ => StatusCode(StatusCodes.Status500InternalServerError, scheduleDto.Error) // fallback para erro desconhecido
+            };
+        }
 
         return Ok(scheduleDto.Value);
     }
     [Authorize]
     [HttpDelete("{id:int}")]
     [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(ScheduleDto))]
-    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Result))]
-    [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Result))]
+    [ProducesResponseType(StatusCodes.Status400BadRequest, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status404NotFound, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized, Type = typeof(Error))]
+    [ProducesResponseType(StatusCodes.Status500InternalServerError, Type = typeof(Error))]
 
-    public async Task<ActionResult<ScheduleDto>> Delete(int id)
+    public async Task<IActionResult> Delete(int id)
     {
+        if (id <= 0)
+        {
+            var error = Result.Failure(Error.BadRequest($"Id inválido: {id}"));
+            return BadRequest(error);
+        }
+
         var result = await _scheduleService.Delete(id);
         if (result.IsFailure)
-            return NotFound(result.Error);
+        {
+            return result.Error.Code switch
+            {
+                "NotFound" => NotFound(result.Error),
+                "BadRequest" => BadRequest(result.Error),
+                "Internal" => StatusCode(StatusCodes.Status500InternalServerError, result.Error),
+                _ => StatusCode(StatusCodes.Status500InternalServerError, result.Error) // fallback para erro desconhecido
+            };
+        }
 
-        var scheduleDto = await _scheduleService.GetById(id);
-        if (scheduleDto.IsFailure)
-            return NotFound(scheduleDto.Error);
+        return Ok(result);
+    }
 
-        return Ok(scheduleDto);
+    private string ErroMoldeState()
+    {
+        var errorMessage = string.Join("; ", ModelState.Values
+                                             .SelectMany(x => x.Errors)
+                                             .Select(x => x.ErrorMessage));
+        return errorMessage;
     }
 }
